migrate track watching from footer to PlayerAudio, fix cache-buster using enum keys
