+++ frontend/src/components/HeroArtist.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import type { SubsonicArtist } from '~/types/subsonicArtist'
-import { artSizes, cacheBustAlbumArt, getCoverArtUrl, onImageError } from '~/logic/common'
+import { artSizes, getCoverArtUrl, onImageError } from '~/logic/common'
 
 const props = defineProps({
   artist: { type: Object as PropType<SubsonicArtist>, required: true },
