+++ frontend/src/logic/common.ts
@@ -70,8 +70,9 @@ export function getCoverArtUrl(musicbrainzId: string, size: number = artSizes.si
 export async function cacheBustAlbumArt(albumId: string) {
   const promises = []
   promises.push(fetch(getCoverArtUrl(albumId), { method: 'POST' }))
-  for (const size of Object.values(artSizes)) {
-    promises.push(fetch(getCoverArtUrl(albumId, Number(size)), { method: 'POST' }))
+  for (const size of Object.values(artSizes).filter(value => typeof value === 'number')) {
+    console.log(`Cache busting album art for album ${albumId} at size ${size}`)
+    promises.push(fetch(getCoverArtUrl(albumId, size), { method: 'POST' }))
   }
   await Promise.all(promises)
 }
