+++ frontend/src/main.ts
@@ -1,8 +1,8 @@
 import type { RouteLocationNormalized, RouteRecordRaw, RouterScrollBehavior } from 'vue-router'
 import { ViteSSG } from 'vite-ssg'
 import { routes } from 'vue-router/auto-routes'
-import { apiKey } from '~/logic/store'
 import { closeSearch } from '~/logic/search'
+import { apiKey } from '~/logic/store'
 import { createEpisodeStoreIfNotExists } from '~/stores/usePodcastStore'
 import App from './App.vue'
